requirements.txt
langchain>=0.1.0
python-dotenv>=1.0.0
pydantic>=2.0.0
llama-cpp-python>=0.2.0
langchain-community>=0.0.10
langchain-ollama>=0.0.1
wikipedia>=1.4.0
pyttsx3
duckduckgo-search
pyaudio
SpeechRecognition
scholarly==1.7.11
colorama
python-telegram-bot>=20.0